# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API para consulta de codigos postales",
        "description": "Busca de codigos postales, información de estados, municipios y sus asentamientos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://codigo-postal.test/api"
        }
    ],
    "paths": {
        "/municipios": {
            "get": {
                "tags": [
                    "municipios"
                ],
                "summary": "Listado de todo los municipios de la republica mexicana",
                "operationId": "5ae14fdfd7846d0ba0620a83fdcc0de9",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Buscar por nombre del municipio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordenar por 'name', '-name'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de registros por consulta",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Si esta activada la paginación, indica a que página desea ir",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Abalá"
                                                    },
                                                    "estado": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Yucatán"
                                                            },
                                                            "abrev": {
                                                                "type": "string",
                                                                "example": "Yuc."
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/municipios/{id}": {
            "get": {
                "tags": [
                    "municipios"
                ],
                "summary": "Obtener datos de un municipio por el id",
                "operationId": "2f8541683e73439eb92fd3109ac5af0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Obtener un municipio por id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 21
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Abalá"
                                                },
                                                "estado": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": "1"
                                                        },
                                                        "nombre": {
                                                            "type": "string",
                                                            "example": "Yucatán"
                                                        },
                                                        "abrev": {
                                                            "type": "string",
                                                            "example": "Yuc."
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recurso Municipality no encontrado, puede que el id no exista en la base de datos"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/municipios/{id}/asentamientos": {
            "get": {
                "tags": [
                    "municipios"
                ],
                "summary": "Obtener los asentamientos de un municipio",
                "operationId": "1987c8841920c97b7cdfad8f7c76ee76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Obtener la lista de asentamientos por municipio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Abalá"
                                                },
                                                "asentamientos": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "cp": {
                                                                "type": "number",
                                                                "example": 97825
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Mukuiche"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recurso Municipality no encontrado, puede que el id no exista en la base de datos"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": "404"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}